Remove most {{{render}}} helpers
These are causing some infinite loops with canjs2.3 and are in most
cases not necessary. It seems like the infinite loop is only caused
when we render a component inside the rendered template.

There are still some cases when we use {{render}} with parameters as a
quick and dirty alternative for components, these will be refactored
into components at a later time.